bannertopdf() templates: Adeed instructions to PDF templates
With the change on the bannertopdf() filter function of the previous
commit the template PDF files can get directly fed into bannertopdf()
and the instructions for adding printer- and job-specific text can be
embedded in the templates.

This commit adds the appropriate instructions to each template file,
to not break PDF integrity to the end. So one can even modify the
instructions without breaking the template itself.

If feeding banner instruction files into bannertopdf() as before, the
embedded instructions get ignored.

Also added files with only the instructions, the data/*-append.txt
files. They are not actually used but only FYI.